+++ documentation/src/assets/popover-icons/contribute.tsx
@@ -1,14 +1,16 @@
 import * as React from "react";
 import { SVGProps } from "react";
 
-const SvgContribute = (props: SVGProps<SVGSVGElement>) => (
+const SvgContribute = ({ className, ...props }: SVGProps<SVGSVGElement>) => (
     <svg
         width="28"
         height="32"
         viewBox="0 0 28 32"
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
+        className={className ? className : undefined}
         {...props}
+        className={props.className ? props.className : undefined}
     >
         <path
             opacity="0.75"