+++ documentation/src/assets/popover-icons/refine-week.tsx
@@ -1,14 +1,16 @@
 import * as React from "react";
 import { SVGProps } from "react";
 
-const SvgRefineWeek = (props: SVGProps<SVGSVGElement>) => (
+const SvgRefineWeek = ({ className, ...props }: SVGProps<SVGSVGElement>) => (
     <svg
         width="32"
         height="32"
         viewBox="0 0 32 32"
         fill="none"
         xmlns="http://www.w3.org/2000/svg"
+        className={className ? className : undefined}
         {...props}
+        className={props.className ? props.className : undefined}
     >
         <g opacity="0.75">
             <path