+++ documentation/src/hooks/use-tutorial-ui-package.ts
@@ -1,0 +1,8 @@
+import React from "react";
+import { TutorialUIPackageContext } from "../context/TutorialUIPackageContext/index";
+
+export const useTutorialUIPackage = () => {
+    const value = React.useContext(TutorialUIPackageContext);
+
+    return value;
+};