+++ documentation/src/theme/DocItem/Layout/index.js
@@ -1,36 +1,16 @@
 import React from "react";
 import clsx from "clsx";
 import { ThemeClassNames } from "@docusaurus/theme-common";
-import { useWindowSize } from "@docusaurus/theme-common";
 import { useDoc } from "@docusaurus/theme-common/internal";
 import DocItemPaginator from "@theme/DocItem/Paginator";
 import DocVersionBanner from "@theme/DocVersionBanner";
 import DocVersionBadge from "@theme/DocVersionBadge";
 import DocItemFooter from "@theme/DocItem/Footer";
-import DocItemTOCMobile from "@theme/DocItem/TOC/Mobile";
-import DocItemTOCDesktop from "@theme/DocItem/TOC/Desktop";
 import DocItemContent from "@theme/DocItem/Content";
 import DocBreadcrumbs from "@theme/DocBreadcrumbs";
 import styles from "./styles.module.css";
-/**
- * Decide if the toc should be rendered, on mobile or desktop viewports
- */
-function useDocTOC() {
-    const { frontMatter, toc } = useDoc();
-    const windowSize = useWindowSize();
-    const hidden = frontMatter.hide_table_of_contents;
-    const canRender = !hidden && toc.length > 0;
-    const mobile = canRender ? <DocItemTOCMobile /> : undefined;
-    const desktop =
-        canRender && (windowSize === "desktop" || windowSize === "ssr") ? (
-            <DocItemTOCDesktop />
-        ) : undefined;
-    return {
-        hidden,
-        mobile,
-        desktop,
-    };
-}
+import { useDocTOCwithTutorial } from "../../../components/tutorial-toc/index";
+import { useCurrentTutorial } from "../../../hooks/use-current-tutorial";
 
 function SwizzleBadge({ className }) {
     return (
@@ -47,7 +27,8 @@ function SwizzleBadge({ className }) {
 }
 
 export default function DocItemLayout({ children }) {
-    const docTOC = useDocTOC();
+    const docTOC = useDocTOCwithTutorial();
+    const tutorial = useCurrentTutorial();
     const { frontMatter, toc } = useDoc();
 
     return (
@@ -55,8 +36,8 @@ export default function DocItemLayout({ children }) {
             <div className={clsx("col", !docTOC.hidden && styles.docItemCol)}>
                 <DocVersionBanner />
                 <div className={styles.docItemContainer}>
-                    <article>
-                        <DocBreadcrumbs />
+                    <article className="doc-article">
+                        {tutorial?.isTutorial ? null : <DocBreadcrumbs />}
                         <div className="flex flex-row gap-1 items-center">
                             <DocVersionBadge />
                             {frontMatter.swizzle && <SwizzleBadge />}
@@ -69,7 +50,9 @@ export default function DocItemLayout({ children }) {
                 </div>
             </div>
             {docTOC.desktop && (
-                <div className="col col--3">{docTOC.desktop}</div>
+                <div className="col col--3 doc--toc-desktop">
+                    {docTOC.desktop}
+                </div>
             )}
         </div>
     );