+++ packages/live-previews/src/scope/headless-inferencer.tsx
@@ -1,0 +1,7 @@
+import * as RefineHeadlessInferencer from "@pankod/refine-inferencer/headless";
+
+const HeadlessInferencerScope = {
+    RefineHeadlessInferencer,
+};
+
+export default HeadlessInferencerScope;