+++ packages/core/CHANGELOG.md
@@ -1,5 +1,79 @@
 # @pankod/refine-core
 
+## 3.94.2
+
+### Patch Changes
+
+-   [#3364](https://github.com/refinedev/refine/pull/3364) [`98a1fbec65a`](https://github.com/refinedev/refine/commit/98a1fbec65abd38da9d6081e04c23b5fe4174acd) Thanks [@aliemir](https://github.com/aliemir)! - Changed `IResourceComponents` from `IResourceContext` to use `React.ComponentType` rather than `React.FunctionComponent` to make it compatible with other types and interfaces.
+
+-   [#3356](https://github.com/refinedev/refine/pull/3356) [`310ebd05990`](https://github.com/refinedev/refine/commit/310ebd05990dd629e64d0a2afcd2b371fe42440f) Thanks [@omeraplak](https://github.com/omeraplak)! - Fixed checkError hook is not called in onError of useCustomMutation
+
+## 3.94.1
+
+### Patch Changes
+
+-   [#3364](https://github.com/refinedev/refine/pull/3364) [`98a1fbec65a`](https://github.com/refinedev/refine/commit/98a1fbec65abd38da9d6081e04c23b5fe4174acd) Thanks [@aliemir](https://github.com/aliemir)! - Changed `IResourceComponents` from `IResourceContext` to use `React.ComponentType` rather than `React.FunctionComponent` to make it compatible with other types and interfaces.
+
+-   [#3356](https://github.com/refinedev/refine/pull/3356) [`310ebd05990`](https://github.com/refinedev/refine/commit/310ebd05990dd629e64d0a2afcd2b371fe42440f) Thanks [@omeraplak](https://github.com/omeraplak)! - Fixed checkError hook is not called in onError of useCustomMutation
+
+## 3.94.0
+
+### Minor Changes
+
+-   [#3335](https://github.com/refinedev/refine/pull/3335) [`ce6acf2b3d4`](https://github.com/refinedev/refine/commit/ce6acf2b3d4c181a87cbdb6c1264fd6e59a504f5) Thanks [@omeraplak](https://github.com/omeraplak)! - feat: `useResource` hook can now also return the current `action`
+
+## 3.93.0
+
+### Minor Changes
+
+-   [#3335](https://github.com/refinedev/refine/pull/3335) [`ce6acf2b3d4`](https://github.com/refinedev/refine/commit/ce6acf2b3d4c181a87cbdb6c1264fd6e59a504f5) Thanks [@omeraplak](https://github.com/omeraplak)! - feat: `useResource` hook can now also return the current `action`
+
+## 3.92.0
+
+### Minor Changes
+
+-   [#3324](https://github.com/refinedev/refine/pull/3324) [`9bfb34749bc`](https://github.com/refinedev/refine/commit/9bfb34749bc8ddaaf80ccffbd0ad6d0a4487309b) Thanks [@aliemir](https://github.com/aliemir)! - Added the ability to pass mutation options to `useMutation` hooks in mutation hooks:
+    -   `useCreate` (data)
+    -   `useUpdate` (data)
+    -   `useDelete` (data)
+    -   `useDeleteMany` (data)
+    -   `useUpdateMany` (data)
+    -   `useCreateMany` (data)
+    -   `useCustomMutation` (data)
+    -   `useLogin` (auth)
+    -   `useLogout` (auth)
+    -   `useRegister` (auth)
+    -   `useForgotPassword` (auth)
+    -   `useUpdatePassword` (auth)
+    -   `useForm` (form)
+
+### Patch Changes
+
+-   [`814eb1009da`](https://github.com/refinedev/refine/commit/814eb1009da53505f7e82f5ed49f2f90e260c316) Thanks [@omeraplak](https://github.com/omeraplak)! - chore: re-exported `@tanstack/react-query`
+
+## 3.91.0
+
+### Minor Changes
+
+-   [#3324](https://github.com/refinedev/refine/pull/3324) [`9bfb34749bc`](https://github.com/refinedev/refine/commit/9bfb34749bc8ddaaf80ccffbd0ad6d0a4487309b) Thanks [@aliemir](https://github.com/aliemir)! - Added the ability to pass mutation options to `useMutation` hooks in mutation hooks:
+    -   `useCreate` (data)
+    -   `useUpdate` (data)
+    -   `useDelete` (data)
+    -   `useDeleteMany` (data)
+    -   `useUpdateMany` (data)
+    -   `useCreateMany` (data)
+    -   `useCustomMutation` (data)
+    -   `useLogin` (auth)
+    -   `useLogout` (auth)
+    -   `useRegister` (auth)
+    -   `useForgotPassword` (auth)
+    -   `useUpdatePassword` (auth)
+    -   `useForm` (form)
+
+### Patch Changes
+
+-   [`814eb1009da`](https://github.com/refinedev/refine/commit/814eb1009da53505f7e82f5ed49f2f90e260c316) Thanks [@omeraplak](https://github.com/omeraplak)! - chore: re-exported `@tanstack/react-query`
+
 ## 3.90.6
 
 ### Patch Changes